Add nonce authentication
This is a bit hacky, but the nonce auth seems worth doing. It
simplifies system configuration/initialization for no need to
rely on sign-in-with-google (3rd party) or noauth (insecure)
to set up the system.  The idea is that by default only nonce
is on to start. User can issue nonce for admin and log in to set
up email and sign in with google to continue logging in. Then user
can enable noauth or not depending on environment.

Ideally, we would not have a user-entered form for nonce auth. It
makes more sense to just paste in a url with the token as a query
param since this is only intended to be relevant for initialization.